Changed Component's "params" to "parameters".
Only required minor changes to Component, and no other classes! A textbook
example of why information hiding is valuable.

Sadly it did mean a whole heap of changes to the Component unit tests, but
that's unavoidable.

I used a general text replace, and all the tests that passed before pass now...
hopefully this won't come back to bite me.