Modified rendering logic.
The TemplateRenderer now renders scalars and hashes using ERB, behaviour
I found myself expecting while working on sample definitions.

To acheive this I had to modify the way the renderer is called, adding a
second, optional parameter that allows you to specify the scope to
render in. This is to allow the InstanceProxy to call the renderer to
render scalars/hashes while maintaining its scope.

I also added a render() function to InstanceProxy, a convenience method
that renders the given target in the scope of the proxy.